+++ lib/src/beamer_router_delegate.dart
@@ -1,27 +1,35 @@
 import 'package:back_button_interceptor/back_button_interceptor.dart';
 import 'package:beamer/beamer.dart';
+import 'package:beamer/src/utils.dart';
 import 'package:flutter/foundation.dart';
-// import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 /// A delegate that is used by the [Router] widget
 /// to build and configure a navigating widget.
-class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
-    with ChangeNotifier, PopNavigatorRouterDelegateMixin<BeamLocation> {
+class BeamerRouterDelegate extends RouterDelegate<Uri>
+    with ChangeNotifier, PopNavigatorRouterDelegateMixin<Uri> {
   BeamerRouterDelegate({
-    @required BeamLocation initialLocation,
+    @required this.beamLocations,
     BeamPage notFoundPage,
+    this.notFoundRedirect,
     this.guards = const <BeamGuard>[],
     this.navigatorObservers = const <NavigatorObserver>[],
   })  : _navigatorKey = GlobalKey<NavigatorState>(),
+        _currentLocation = beamLocations[0]..prepare(),
         notFoundPage = notFoundPage ?? BeamPage(child: Container()) {
-    // _beamHistory.add(initialLocation..prepare());
-    _currentLocation = initialLocation..prepare();
-    BackButtonInterceptor.add(backInterceptor, name: 'BeamerInterceptor');
+    if (!kIsWeb) {
+      BackButtonInterceptor.add(backInterceptor, name: 'BeamerInterceptor');
+    }
   }
+
+  /// List of all [BeamLocation]s that this router handles.
+  final List<BeamLocation> beamLocations;
 
   /// Page to show when no [BeamLocation] supports the incoming URI.
   final BeamPage notFoundPage;
+
+  /// [BeamLocation] to redirect to when no [BeamLocation] supports the incoming URI.
+  final BeamLocation notFoundRedirect;
 
   /// Guards that will be executing [check] on [currentLocation] candidate.
   ///
@@ -43,8 +51,6 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   BeamLocation _currentLocation;
 
   /// Access the current [BeamLocation].
-  ///
-  /// The same thing as [currentConfiguration], but with more familiar name.
   ///
   /// Can be useful in:
   ///
@@ -68,7 +74,11 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   List<BeamPage> get currentPages => _currentPages;
 
   /// Whether to implicitly [beamBack] instead of default pop
-  bool _beamBackOnPop;
+  bool _beamBackOnPop = false;
+
+  /// Whether all the pages from location are stacked.
+  /// If not (`false`), just the last page is taken.
+  bool _stacked = true;
 
   /// Beams to `location`.
   ///
@@ -78,14 +88,53 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   /// 2. updates [currentLocation]
   /// 3. notifies that the [Navigator] should be rebuilt
   ///
-  /// if `beamBackOnPop` is set to `true`, default pop action on the newly
+  /// If `beamBackOnPop` is set to `true`, default pop action on the newly
   /// beamed location will triger `beamBack` instead.
-  void beamTo(BeamLocation location, {bool beamBackOnPop = false}) {
+  /// If `stacked` is set to `false`, only the location's last page will be shown.
+  void beamTo(
+    BeamLocation location, {
+    bool beamBackOnPop = false,
+    bool stacked = true,
+  }) {
     _beamBackOnPop = beamBackOnPop;
+    _stacked = stacked;
     _beamHistory.add(location..prepare());
     _currentLocation = _beamHistory.last;
     notifyListeners();
   }
+
+  /// Beams to [BeamLocation] that handles `uri`. See [beamTo].
+  ///
+  /// For example
+  ///
+  /// ```dart
+  /// Beamer.of(context).beamToNamed(
+  ///   '/user/1/transactions?perPage=10',
+  ///   data: {'beenHereBefore': true},
+  /// );
+  /// ```
+  ///
+  /// `data` can be used to pass any data throught the location.
+  /// See [BeamLocation.data].
+  void beamToNamed(
+    String uri, {
+    Map<String, dynamic> data = const <String, dynamic>{},
+    bool beamBackOnPop = false,
+    bool stacked = true,
+  }) {
+    final location = Utils.chooseBeamLocation(Uri.parse(uri), beamLocations);
+    location.data = data;
+    beamTo(location, beamBackOnPop: beamBackOnPop);
+  }
+
+  /// Whether it is possible to [beamBack],
+  /// i.e. there is more than 1 location in [beamHistory].
+  bool get canBeamBack => _beamHistory.length > 1;
+
+  /// What is the location to which [beamBack] will lead.
+  /// If there is none, returns null.
+  BeamLocation get beamBackLocation =>
+      canBeamBack ? _beamHistory[_beamHistory.length - 2] : null;
 
   /// Beams to previous location in [beamHistory]
   /// and **removes** the last location from history.
@@ -95,7 +144,8 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   /// Returns the success, whether the [currentLocation] was changed.
   bool beamBack() {
     _beamBackOnPop = false;
-    if (_beamHistory.length == 1) {
+    _stacked = true;
+    if (!canBeamBack) {
       return false;
     }
     _beamHistory.removeLast();
@@ -108,15 +158,19 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   ///
   /// See [BeamLocation.update] for details.
   ///
-  /// Note that [_beamBackOnPop] will be reset to `false`.
+  /// If `beamBackOnPop` is not specified, [_beamBackOnPop] will be reset to `false`.
+  /// If `stacked` is not specified, [_stacked] will remain as it was.
   void updateCurrentLocation({
-    @required String pathBlueprint,
+    String pathBlueprint,
     Map<String, String> pathParameters = const <String, String>{},
     Map<String, String> queryParameters = const <String, String>{},
     Map<String, dynamic> data = const <String, dynamic>{},
     bool rewriteParameters = false,
+    bool beamBackOnPop,
+    bool stacked,
   }) {
-    _beamBackOnPop = false;
+    _beamBackOnPop = beamBackOnPop ?? false;
+    _stacked = stacked ?? _stacked;
     _currentLocation.update(
       pathBlueprint: pathBlueprint,
       pathParameters: pathParameters,
@@ -129,7 +183,7 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   }
 
   @override
-  BeamLocation get currentConfiguration => _currentLocation;
+  Uri get currentConfiguration => _currentLocation.uri;
 
   @override
   GlobalKey<NavigatorState> get navigatorKey => _navigatorKey;
@@ -138,11 +192,16 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   Widget build(BuildContext context) {
     final BeamGuard guard = _guardCheck(context, _currentLocation);
     if (guard?.beamTo != null) {
-      beamTo(guard.beamTo(context));
+      _beamHistory.add(guard.beamTo(context)..prepare());
+      _currentLocation = _beamHistory.last;
+    } else if ((_currentLocation is NotFound) && notFoundRedirect != null) {
+      _currentLocation = notFoundRedirect..prepare();
     }
     final navigator = Builder(
       builder: (context) {
-        _currentPages = _currentLocation.pagesBuilder(context);
+        _currentPages = _stacked
+            ? _currentLocation.pagesBuilder(context)
+            : [_currentLocation.pagesBuilder(context).last];
         return Navigator(
           key: navigatorKey,
           observers: navigatorObservers,
@@ -173,8 +232,8 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   }
 
   @override
-  SynchronousFuture<void> setNewRoutePath(BeamLocation location) {
-    beamTo(location);
+  SynchronousFuture<void> setNewRoutePath(Uri uri) {
+    beamTo(Utils.chooseBeamLocation(uri, beamLocations));
     return SynchronousFuture(null);
   }
 
@@ -193,11 +252,13 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
   BeamGuard _guardCheck(BuildContext context, BeamLocation location) {
     for (var guard in guards) {
       if (guard.shouldGuard(location) && !guard.check(context, location)) {
+        guard.onCheckFailed?.call(context, location);
         return guard;
       }
     }
     for (var guard in location.guards) {
       if (guard.shouldGuard(location) && !guard.check(context, location)) {
+        guard.onCheckFailed?.call(context, location);
         return guard;
       }
     }
@@ -206,15 +267,16 @@ class BeamerRouterDelegate extends RouterDelegate<BeamLocation>
 
   @override
   void dispose() {
-    BackButtonInterceptor.removeByName('BeamerInterceptor');
+    if (!kIsWeb) {
+      BackButtonInterceptor.removeByName('BeamerInterceptor');
+    }
     super.dispose();
   }
 
   bool backInterceptor(bool stopDefaultButtonEvent, RouteInfo info) {
-    var canPop = false;
-    if (_beamHistory.length > 1 && currentPages.length == 1) {
-      canPop = beamBack();
+    if (_currentPages.length == 1) {
+      return beamBack();
     }
-    return canPop;
+    return false;
   }
 }