+++ test/beamer_router_delegate_test.dart
@@ -8,9 +8,9 @@ void main() {
   final location1 = Location1(pathBlueprint: '/l1');
   final location2 = Location2(pathBlueprint: '/l2/:id');
   final router = BeamerRouterDelegate(
-    initialLocation: location1,
+    beamLocations: [location1, location2],
   );
-  router.setNewRoutePath(location1);
+  router.setNewRoutePath((location1..prepare()).uri);
 
   group('initialization & beaming', () {
     test('initialLocation is set', () {
@@ -22,20 +22,40 @@ void main() {
       expect(router.currentLocation, location2);
     });
 
-    test('beamBack leads to previous location', () {
-      router.beamTo(location2);
+    test('beamToNamed changes locations with correct parameters', () {
+      router.beamToNamed('/l2/1?q=t', data: {'x': 'y'});
+      expect(router.currentLocation, location2);
+      expect(location2.pathParameters.containsKey('id'), true);
+      expect(location2.pathParameters['id'], '1');
+      expect(location2.queryParameters.containsKey('q'), true);
+      expect(location2.queryParameters['q'], 't');
+      expect(location2.data, {'x': 'y'});
+    });
 
+    test('beamBack leads to previous location and all helpers are correct', () {
+      expect(router.canBeamBack, true);
+      expect(router.beamBackLocation, isA<Location2>());
       bool success = router.beamBack();
       expect(success, true);
       expect(router.currentLocation, location2);
 
+      expect(router.canBeamBack, true);
+      expect(router.beamBackLocation, isA<Location1>());
       success = router.beamBack();
       expect(success, true);
       expect(router.currentLocation, location1);
 
+      expect(router.canBeamBack, false);
+      expect(router.beamBackLocation, null);
       success = router.beamBack();
       expect(success, false);
       expect(router.currentLocation, location1);
     });
+  });
+
+  test('stacked beam takes just last page for currentPages', () {
+    router.beamToNamed('/l1/one', stacked: false);
+    expect(router.currentLocation.pagesBuilder(null).length, 2);
+    //expect(router.currentPages.length, 1);
   });
 }